+++ src/biz/collab/utils.rs
@@ -297,3 +297,60 @@ pub fn collab_from_doc_state(doc_state: Vec<u8>, object_id: &str) -> Result<Coll
   .map_err(|e| AppError::Unhandled(e.to_string()))?;
   Ok(collab)
 }
+
+/// Base on values given by [cell_value_by_id], write to fields of DatabaseRowBody.
+/// Returns encoded collab updates to the database row
+pub async fn write_to_database_row(
+  db_collab: &Collab,
+  db_body: &DatabaseBody,
+  db_row_collab: &mut Collab,
+  db_row_body: &DatabaseRowBody,
+  cell_value_by_id: HashMap<String, serde_json::Value>,
+  modified_ts: i64,
+) -> Result<Vec<u8>, AppError> {
+  let all_fields = db_body.fields.get_all_fields(&db_collab.transact());
+  let field_by_id = all_fields.iter().fold(HashMap::new(), |mut acc, field| {
+    acc.insert(field.id.clone(), field.clone());
+    acc
+  });
+  let type_option_reader_by_id = type_option_writer_by_id(&all_fields);
+  let field_by_name = field_by_name_uniq(all_fields);
+
+  let db_row_collab_updates = {
+    let mut yrs_txn = db_row_collab.transact_mut();
+    // set last_modified
+    db_row_body.update(&mut yrs_txn, |row_update| {
+      row_update.set_last_modified(modified_ts);
+    });
+
+    // for each field given by user input, overwrite existing data
+    for (id, serde_val) in cell_value_by_id {
+      let field = match field_by_id.get(&id) {
+        Some(f) => f,
+        // try use field name if id not found
+        None => match field_by_name.get(&id) {
+          Some(f) => f,
+          None => {
+            tracing::warn!("Failed to get field by id or name for field: {}", id);
+            continue;
+          },
+        },
+      };
+      let cell_writer = match type_option_reader_by_id.get(&field.id) {
+        Some(cell_writer) => cell_writer,
+        None => {
+          tracing::error!("Failed to get type option writer for field: {}", field.id);
+          continue;
+        },
+      };
+      let new_cell: Cell = cell_writer.convert_json_to_cell(serde_val);
+      db_row_body.update(&mut yrs_txn, |row_update| {
+        row_update.update_cells(|cells_update| {
+          cells_update.insert_cell(&field.id, new_cell);
+        });
+      });
+    }
+    yrs_txn.encode_update_v1()
+  };
+  Ok(db_row_collab_updates)
+}
